requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dependency-injector==4.41.0
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.92.0
fastapi-utils==0.2.1
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
h11==0.14.0
hiredis==2.2.2
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pyparsing==3.0.9
pytest==7.2.1
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
